--- public/database2.json
@@ -1,7 +1,0 @@
-{
-    "data": [
-        { "name": "meizu", "price": 108 },
-        { "name": "asus", "price": 1500 },
-        { "name": "acer", "price": 1350 }
-      ]
-}
