--- public/database3.json
@@ -1,7 +1,0 @@
-{
-    "data": [
-        { "name": "motorola", "price": 980 },
-        { "name": "sony", "price": 450 },
-        { "name": "lenovo", "price": 378 }
-      ]
-}
