--- public/database4.json
@@ -1,7 +1,0 @@
-{
-    "data": [
-        { "name": "braun", "price": 46 },
-        { "name": "toshiba", "price": 578 },
-        { "name": "vitek", "price": 4500 }
-      ]
-}
