+++ public/index.html
@@ -1,32 +1,30 @@
 <!DOCTYPE html>
 <html lang="ru">
 <head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Alibaba</title>
-    <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@500;700&display=swap" rel="stylesheet">
-    <link rel="stylesheet" href="style.css">
-    <script src="script.js" defer></script>
+  <meta charset="UTF-8">
+  <meta name="viewport" content="width=device-width, initial-scale=1.0">
+  <title>My mini-market</title>
+  <link href="https://fonts.googleapis.com/css2?family=Roboto:wght@500;700&display=swap" rel="stylesheet">
+  <!-- <link rel="stylesheet" href="styles/style.css"> -->
+  <script src="js/shop.js" defer></script>
 </head>
 <body>
-    <header>
-        <div class="basket h-50">
-            <div class="basket-button-wrap"></div>
-            <div class="basket-list"></div>
-            <div class="basket-sum"></div>
-        </div>
-    </header>
-    <aside>
-    </aside>
-    <main>
-        <div class="goods">
-            <div class="goods-list"></div>
-            <div class="goods-button"></div>
-        </div>
-        
-        
-    </main>
-    
+  <header>
+    <span>Урок 5</span>
+    <div class="basket h-50">
+      <div class="basket-button-wrap"></div>
+      <div class="basket-list"></div>
+      <div class="basket-sum"></div>
+    </div>
+  </header>
+  <aside>
+  </aside>
+  <main>
+    <div class="goods">
+      <div class="goods-list"></div>
+      <div class="goods-button"></div>
+    </div>
+  </main>
 </body>
 </html>
