+++ src/style.css
@@ -1,0 +1,198 @@
+* {
+  margin: 0;
+  padding: 0;
+}
+body {
+  display: flex;
+  font-family: 'Roboto', sans-serif;
+  color: rgb(19, 19, 19);
+  background-color: rgb(237, 237, 237);
+}
+header {
+  width: 100%;
+  height: 50px;
+  padding-right: 30px;
+  padding-left: 30px;
+  box-sizing: border-box;
+  position: fixed;
+  display: flex;
+  justify-content: space-between;
+  top: 0;
+  background-color:rgb(65, 65, 65);
+}
+header span {
+  margin-top: auto;
+  margin-bottom: auto;
+  color: #ffffff;
+}
+aside {
+  width: 300px;
+  height: 100vh;
+  padding-top: 50px;
+  box-sizing: border-box;
+  display: flex;
+  background-color: rgb(215, 215, 215);
+}
+
+.form-wrap {
+  width: 250px;
+  margin: auto;
+  display: flex;
+  flex-direction: column;
+}
+.form-wrap>form {
+  width: 100%;
+  display: flex;
+  box-sizing: border-box;
+  flex-direction: column;
+}
+.form-wrap>form input {
+  height: 25px;
+  margin: 10px;
+  padding: 7px;
+  box-sizing: border-box;
+}
+.form-wrap>button {
+  margin: auto;
+}
+
+main {
+  width: calc(100% - 300px);
+  height: 100vh;
+  padding-top: 50px;
+  box-sizing: border-box;
+  display: flex;
+}
+.goods-list {
+  width: 100%;
+  height: calc(100% - 70px);
+  overflow-y: scroll;
+  box-sizing: border-box;
+  justify-content: center;
+  display: flex;
+  flex-wrap: wrap;
+}
+.goods {
+  width: 100%;
+  height: 100%;
+  box-sizing: border-box;
+  justify-content: center;
+  display: flex;
+  flex-direction: column;
+}
+.goods-button {
+  width: 100%;
+  height: 70px;
+  align-items: center;
+  text-align: center;
+  box-sizing: border-box;
+  padding: 15px;
+  background-color:rgb(65, 65, 65);
+}
+.basket {
+  width: 500px;
+  height: 300px;
+  margin-left: auto;
+  box-sizing: border-box;
+  display: flex;
+  flex-direction: column;
+}
+.basket-button-wrap {
+  height: 50px;
+  display: flex;
+  margin-left: auto;
+}
+.basket-button-wrap>button {
+  margin-top: auto;
+  margin-bottom: auto;
+  margin-left: auto;
+
+}
+.basket-list {
+  width: 100%;
+  height: calc(100% - 50px);
+  margin-top: 6px;
+  overflow-y: scroll;
+  box-sizing: border-box;
+  display: none;
+  background-color:rgb(75, 75, 75);
+}
+.basket-sum {
+  width: 100%;
+  height: 35px;
+  text-align: center;
+  box-sizing: border-box;
+  padding: 5px;
+  display: none;
+  background-color:rgb(65, 65, 65);
+}
+.basket-block {
+  display: block !important;
+}
+.h-50 {
+  height: 50px !important;
+}
+.goods-list_item {
+  width: 250px;
+  height: 200px;
+  margin: 10px;
+  justify-content: center;
+  align-items: center;
+  text-transform: uppercase;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  border: 1px solid rgb(57, 57, 57);
+  background-color: rgb(237, 237, 237);
+}
+.basket-list_item {
+  height: 35px;
+  margin: 5px;
+  padding: 5px;
+  text-transform: uppercase;
+  display: flex;
+  /* justify-content: space-between; */
+  box-sizing: border-box;
+  border: 1px solid rgb(57, 57, 57);
+  background-color: rgb(237, 237, 237);
+}
+.basket-list_item > :nth-child(1) {
+  margin-right: 20px;
+}
+.basket-list_item > :nth-child(3) {
+  margin-left: auto;
+}
+.basket-list_item > :nth-child(4) {
+  margin-left: 10px;
+  margin-right: 10px;
+}
+.basket-list_item > :nth-child(5) {
+  margin-right: 10px;
+}
+
+
+.list_item {
+  width: 150px;
+  height: 200px;
+  margin: 10px;
+  justify-content: center;
+  align-items: center;
+  text-transform: uppercase;
+  display: flex;
+  flex-direction: column;
+  box-sizing: border-box;
+  background-color: rgb(54, 69, 138);
+}
+
+.btn {
+  background: rgb(43, 43, 43);
+  border: none !important;
+  padding: 10px;
+  cursor: pointer;
+  border-radius: 5px;
+  color: #ffffff;
+}
+.btn:hover {
+  background: rgb(33, 33, 33);
+  border: none !important;
+}
